explorer-api: move mix node client operations into a package
