+++ explorer-api/src/country_statistics/mod.rs
@@ -1,12 +1,11 @@
 use isocountry::CountryCode;
 use log::{info, trace, warn};
-use mixnet_contract::MixNodeBond;
 use reqwest::Error as ReqwestError;
-use validator_client::Config;
 
 use models::GeoLocation;
 
 use crate::country_statistics::country_nodes_distribution::CountryNodesDistribution;
+use crate::mix_nodes::retrieve_mixnodes;
 use crate::state::ExplorerApiStateContext;
 
 pub mod country_nodes_distribution;
@@ -102,23 +101,4 @@ async fn locate(ip: &str) -> Result<GeoLocation, ReqwestError> {
     let response = reqwest::get(format!("{}{}", crate::GEO_IP_SERVICE, ip)).await?;
     let location = response.json::<GeoLocation>().await?;
     Ok(location)
-}
-
-async fn retrieve_mixnodes() -> Vec<MixNodeBond> {
-    let client = new_validator_client();
-
-    info!("About to retrieve mixnode bonds...");
-
-    let bonds: Vec<MixNodeBond> = match client.get_cached_mix_nodes().await {
-        Ok(result) => result,
-        Err(e) => panic!("Unable to retrieve mixnode bonds: {:?}", e),
-    };
-    info!("Fetched {} mixnode bonds", bonds.len());
-    bonds
-}
-
-// TODO: inject constants
-fn new_validator_client() -> validator_client::Client {
-    let config = Config::new(vec![crate::VALIDATOR_API.to_string()], crate::CONTRACT);
-    validator_client::Client::new(config)
 }
