+++ explorer-api/src/main.rs
@@ -7,6 +7,7 @@ use log::info;
 
 mod country_statistics;
 mod http;
+mod mix_nodes;
 mod ping;
 mod state;
